+++ source/TestCommon/source/FunctionApp.TestCommon.Tests/Integration/Configuration/IntegrationTestConfigurationTests.cs
@@ -1,0 +1,48 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Energinet.DataHub.Core.FunctionApp.TestCommon.Configuration;
+using Energinet.DataHub.Core.TestCommon;
+using FluentAssertions;
+using Xunit;
+
+namespace Energinet.DataHub.Core.FunctionApp.TestCommon.Tests.Integration.Configuration
+{
+    public class IntegrationTestConfigurationTests : TestBase<IntegrationTestConfiguration>
+    {
+        [Fact]
+        public void Given_IdentityHasAccess_When_ServiceBusConnectionString_Then_HasValue()
+        {
+            // Arrange
+
+            // Act
+            var actualValue = Sut.ServiceBusConnectionString;
+
+            // Assert
+            actualValue.Should().NotBeNullOrEmpty();
+        }
+
+        [Fact]
+        public void Given_IdentityHasAccess_When_ApplicationInsightsInstrumentationKey_Then_HasValue()
+        {
+            // Arrange
+
+            // Act
+            var actualValue = Sut.ApplicationInsightsInstrumentationKey;
+
+            // Assert
+            actualValue.Should().NotBeNullOrEmpty();
+        }
+    }
+}