+++ source/TestCommon/source/FunctionApp.TestCommon/FunctionAppHost/FunctionAppHostSettings.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System.Collections.Generic;
+
 namespace Energinet.DataHub.Core.FunctionApp.TestCommon.FunctionAppHost
 {
     /// <summary>
@@ -46,11 +48,13 @@ namespace Energinet.DataHub.Core.FunctionApp.TestCommon.FunctionAppHost
         /// functions output (logs) in the console for e.g. debugging issues.
         /// </summary>
         public bool UseShellExecute { get; set; }
+            = false;
 
         /// <summary>
         /// Port on which the host will listen.
         /// </summary>
         public int Port { get; set; }
+            = 7001;
 
         /// <summary>
         /// After starting the host this is the max. time we wait before we expect the
@@ -59,5 +63,19 @@ namespace Energinet.DataHub.Core.FunctionApp.TestCommon.FunctionAppHost
         /// If the host is not ready within expected time, an exception is thrown.
         /// </summary>
         public int MaxWaitSeconds { get; set; }
+            = 120;
+
+        /// <summary>
+        /// A space separated list of functions to load. If empty all functions will be loaded.
+        /// </summary>
+        public string Functions { get; set; }
+            = string.Empty;
+
+        /// <summary>
+        /// Only support if <see cref="UseShellExecute"/> is "false".
+        /// A dictionary of environment variable that will we set for the Function App host process.
+        /// </summary>
+        public IDictionary<string, string> ProcessEnvironmentVariables { get; set; }
+            = new Dictionary<string, string>();
     }
 }