+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -1,5 +1,16 @@
 # TestCommon Release notes
 
+## Version 1.1.0
+
+- Extended control of starting an Azure Function App through use of the `FunctionAppHostManager`:
+  - Specify which functions to load using `FunctionAppHostSettings.Functions`.
+  - Set environment variables for the process using `FunctionAppHostSettings.ProcessEnvironmentVariables`.
+  - Write host process id to output when started, so we know which process to attach debugger to if needed.
+- Updated default values for `FunctionAppHostSettings` to reduce the necessary settings in  `functionapphost.settings.json`.
+- Implemented `FunctionAppHostManagerExtensions`:
+  - `CheckIfFunctionWasExecuted`
+  - `TriggerFunctionAsync`
+
 ## Version 1.0.1
 
 - Ensure `AzuriteManager` does not fail when Azure Storage Emulator is not installed on machine.